+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [vNext]
 
-- ...
+### Changed
+
+- Updated Substrate to polkadot-v0.9.27
 
 ## [4.4.0]