+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [vNext]
 
+## [4.5.0]
+
 ### Changed
 
 - Updated Substrate to polkadot-v0.9.27